+++ src/test/kotlin/blackjack/TestCards.kt
@@ -1,4 +1,8 @@
-package blackjack.model
+package blackjack
+
+import blackjack.model.Card
+import blackjack.model.Rank
+import blackjack.model.Suit
 
 object TestCards {
     val Ace = Card.of(Suit.DIAMONDS, Rank.ACE)
