+++ src/test/kotlin/blackjack/model/HandTest.kt
@@ -1,5 +1,6 @@
 package blackjack.model
 
+import blackjack.TestCards
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
