+++ src/test/kotlin/blackjack/states/FirstTurnTest.kt
@@ -1,6 +1,6 @@
 package blackjack.states
 
-import blackjack.model.TestCards
+import blackjack.TestCards
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.assertTrue
